ICOMP-403 Remove remove-baseline and clean-up-baseline commands (plan is to add them back later) and fix SA warnings

+++ tests/PhpCodeSnifferRunner/Report/ReportFactory.php
@@ -23,10 +23,10 @@ class ReportFactory
         $fileReportsByFilename = $fileReportsByFilenameOrDefault !== null ? $fileReportsByFilenameOrDefault : [];
         return new Report(
             $totals !== null ? $totals : TotalsFactory::create(
-                array_sum(array_map(static function (FileReport $fileReport): int {
+                (int) array_sum(array_map(static function (FileReport $fileReport): int {
                     return $fileReport->getErrorCount();
                 }, $fileReportsByFilename)),
-                array_sum(array_map(static function (FileReport $fileReport): int {
+                (int) array_sum(array_map(static function (FileReport $fileReport): int {
                     return $fileReport->getWarningCount();
                 }, $fileReportsByFilename)),
             ),
